# backend/openapi/v1.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "CPX Secreto API",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "operationId": "getDashboard",
                "responses": {
                    "200": {
                        "description": "The authenticated user home dashboard.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Dashboard"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication is required.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Dashboard": {
                "required": [
                    "featuredEdition",
                    "groups"
                ],
                "properties": {
                    "featuredEdition": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/DashboardEdition"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "groups": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/DashboardGroup"
                        }
                    }
                },
                "type": "object"
            },
            "DashboardEdition": {
                "required": [
                    "groupId",
                    "groupName",
                    "editionId",
                    "editionName",
                    "status",
                    "budgetCents",
                    "currency",
                    "eventDate",
                    "isAdmin",
                    "isParticipant",
                    "participantCount",
                    "wishCount",
                    "assignmentAvailable"
                ],
                "properties": {
                    "budgetCents": {
                        "oneOf": [
                            {
                                "type": "integer",
                                "minimum": 0
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "eventDate": {
                        "oneOf": [
                            {
                                "type": "string",
                                "format": "date"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "groupId": {
                        "type": "integer",
                        "example": 1
                    },
                    "groupName": {
                        "type": "string",
                        "example": "Família Silva"
                    },
                    "editionId": {
                        "type": "integer",
                        "example": 4
                    },
                    "editionName": {
                        "type": "string",
                        "example": "Natal 2026"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "draft",
                            "open",
                            "drawn",
                            "revealed"
                        ]
                    },
                    "currency": {
                        "type": "string",
                        "example": "BRL"
                    },
                    "isAdmin": {
                        "type": "boolean"
                    },
                    "isParticipant": {
                        "type": "boolean"
                    },
                    "participantCount": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "wishCount": {
                        "type": "integer",
                        "minimum": 0
                    },
                    "assignmentAvailable": {
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "DashboardGroup": {
                "required": [
                    "id",
                    "name",
                    "memberCount",
                    "currentEditionId",
                    "currentEditionStatus"
                ],
                "properties": {
                    "currentEditionId": {
                        "oneOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "currentEditionStatus": {
                        "oneOf": [
                            {
                                "type": "string",
                                "enum": [
                                    "draft",
                                    "open",
                                    "drawn",
                                    "revealed"
                                ]
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Família Silva"
                    },
                    "memberCount": {
                        "type": "integer",
                        "example": 8,
                        "minimum": 1
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Não autenticado."
                    },
                    "errors": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}